minor updates to selling of dorm

+++ src/com/endicott/edu/simulators/DormManager.java
@@ -166,7 +166,7 @@ public class DormManager {
             name = d.getName();
             totalBuildCost = d.getTotalBuildCost();
             //takes 20% of the build cost to refund back to the college.
-            refund = totalBuildCost/20;
+            refund = (int)(totalBuildCost/20);
             if(name.equals(dormName)){
                 dorms.remove(d);
                 Accountant.studentIncome(runId, dormName + "has been sold.", refund);
